feat: atualização dos valores de ID dos documentos, pois, não é necessário ser Long, Integer já serve para o propósito

+++ service/local-layer/register-service/src/main/java/propagandus/registerservice/core/repositorys/AttentionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import propagandus.registerservice.core.documents.Attention;
 
 @Repository
-public interface AttentionRepository extends MongoRepository<Attention,Long> {
+public interface AttentionRepository extends MongoRepository<Attention,Integer> {
 
 }
