+++ service/local-layer/register-service/src/main/java/propagandus/registerservice/core/repositorys/ReactionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import propagandus.registerservice.core.documents.Reaction;
 
 @Repository
-public interface ReactionRepository extends MongoRepository<Reaction,Long> {
+public interface ReactionRepository extends MongoRepository<Reaction,Integer> {
 
 }
